refactor: organize project structure for FastAPI
