+++ markdown_processing/cleaner.py
@@ -1,7 +1,7 @@
 # markdown_processing/cleaner.py
 from llama_index.core import Settings
 from llama_index.core.prompts import PromptTemplate
-from utility import split_markdown_by_heading, run_llm_process, write_debug_chunks
+from .utils import split_markdown_by_heading, run_llm_process, write_debug_chunks
 from pathlib import Path
 from config import max_chars_per_chunk, DEBUG_CLEAN_DIR
